+++ src/main/java/pl/thedeem/intellij/dql/psi/elements/DQLParametersOwner.java
@@ -1,10 +1,10 @@
 package pl.thedeem.intellij.dql.psi.elements;
 
+import com.intellij.psi.PsiElement;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import pl.thedeem.intellij.dql.definition.model.MappedParameter;
 import pl.thedeem.intellij.dql.definition.model.Parameter;
-import pl.thedeem.intellij.dql.psi.DQLExpression;
 
 import java.util.Collection;
 import java.util.List;
@@ -18,5 +18,5 @@ public interface DQLParametersOwner {
 
     @NotNull Collection<Parameter> getMissingParameters();
 
-    @Nullable MappedParameter getParameter(@NotNull DQLExpression parameter);
+    @Nullable MappedParameter getParameter(@NotNull PsiElement parameter);
 }