+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Adding support for scientific notation numbers in DQL
 - Adding support for calculating the expressions data types for supported operators (if possible, for example: `long` +
   `long`)
+- Added support for fields named via `alias:` parameter. Added intention to replace `alias` with
+  assigned expression (`=`)
 - Replacing manual inspections with automatically generated based on the DQL JSON definition
 
 ## [1.2.0] - 2025-12-17